fixme ScottyException handling still not working
